pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drive-vertex-cli"
version = "0.1.0"
description = "Python library and CLI for indexing Google Drive content with Vertex AI embeddings and answering questions with Gemini."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-api-python-client>=2.170.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.2",
  "google-genai>=1.30.0",
  "numpy>=2.2.0",
  "openpyxl>=3.1.5",
  "pypdf>=5.4.0",
  "python-docx>=1.1.2",
  "python-dotenv>=1.0.1",
  "python-pptx>=1.0.2",
  "rich>=13.9.0",
  "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
]

[project.scripts]
drive-vertex = "drive_vertex_cli.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
